Add more compiler warnings.

+++ Fractron/DFMandelbrot.m
@@ -15,7 +15,7 @@
 
 BRU_DEFAULT_INIT_UNAVAILABLE_IMPL
 
-- (nonnull instancetype)initWithIterations:(NSInteger)iterations
+- (nonnull instancetype)initWithIterations:(NSUInteger)iterations
                                 dimensions:(NSSize)dimensions
                                     region:(NSRect)region
 {
@@ -30,29 +30,29 @@ BRU_DEFAULT_INIT_UNAVAILABLE_IMPL
 }
 
 + (NSInteger)calculateIterationsForPoint:(NSPoint)c
-                              iterations:(NSInteger)iterations
+                              iterations:(NSUInteger)iterations
 {
     NSPoint z = c;
-    for (NSInteger i = 0; i < iterations; i++) {
+    for (NSUInteger i = 0; i < iterations; i++) {
         NSPoint old = z;
         z.x = (old.x * old.x) - (old.y * old.y) + c.x;
         z.y = (2 * old.x * old.y) + c.y;
         if (((z.x * z.x) + (z.y * z.y)) > 4.0) {
-            return i;
+            return (NSInteger)i;
         }
     }
     return -1;
 }
 
-+ (void)generateRow:(NSInteger)y
++ (void)generateRow:(NSUInteger)y
            fractalY:(CGFloat)fractalY
            fractalX:(CGFloat)fractalX
              x_skip:(CGFloat)x_skip
              buffer:(uint8_t*)p
-              width:(NSInteger)width
-         iterations:(NSInteger)iterations
+              width:(NSUInteger)width
+         iterations:(NSUInteger)iterations
 {
-    for (NSInteger x = 0; x < width; x++) {
+    for (NSUInteger x = 0; x < width; x++) {
         NSPoint fractalPoint = NSMakePoint(fractalX + (x_skip * x), fractalY);
         NSInteger escape = [DFMandelbrot calculateIterationsForPoint:fractalPoint
                                                           iterations:iterations];
@@ -75,14 +75,14 @@ BRU_DEFAULT_INIT_UNAVAILABLE_IMPL
             return;
         }
     
-        NSMutableData *data = [[NSMutableData alloc] initWithLength: self.dimensions.width * self.dimensions.height];
+        NSMutableData *data = [[NSMutableData alloc] initWithLength:(NSUInteger)(self.dimensions.width * self.dimensions.height)];
         BRUAssert(data, @"Failed to allocate data!"); // docs imply this never fails
         uint8_t *p = (uint8_t*)data.bytes;
         
         double x_skip = self.region.size.width / self.dimensions.width;
         double y_skip = self.region.size.height / self.dimensions.height;
         
-        for (NSInteger y = 0; y < self.dimensions.height; y++) {
+        for (NSUInteger y = 0; y < self.dimensions.height; y++) {
             
             dispatch_async(dispatch_get_global_queue(QOS_CLASS_BACKGROUND, 0), ^() {
                 BRU_strongify(self);
@@ -95,7 +95,7 @@ BRU_DEFAULT_INIT_UNAVAILABLE_IMPL
                                  fractalX:self.region.origin.x
                                    x_skip:x_skip
                                    buffer:p
-                                    width:self.dimensions.width
+                                    width:(NSUInteger)self.dimensions.width
                                iterations:self.iterations];
                 
                 callback(self, data);
